pyproject.toml
[project]
name = "shop-management-fastapi"
version = "0.1.0"
description = "Small store management backend (inventory, invoicing) built with FastAPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0,<0.110.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "alembic>=1.11.0",
    "pydantic<2.0.0",
    "fastapi-jwt-auth>=0.5.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "flake8>=6.1.0",
    "coverage>=7.3.0",
    "httpx>=0.25.0,<0.28.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.flake8]
max-line-length = 127
max-complexity = 10
exclude = ["migrations/*", "__pycache__/*", ".venv/*"]
